pyproject.toml
[tool.poetry]
name = "imgdl"
version = "2.1.0-beta.1"
description = "Bulk image downloader from a list of urls"
authors = ["Felipe Aguirre Martinez <felipeam86@gmail.com>"]
readme = "README.rst"
homepage = "https://github.com/felipeam86/imagedownloader"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Pillow = "^9.2.0"
requests = "^2.28.1"
tqdm = "^4.64.1"
python-json-logger = "^2.0.4"
pydantic = "^1.10.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
jupyter = "^1.0.0"
black = "^22.10.0"
isort = "^5.10.1"
pytest-cov = "^4.0.0"


[tool.poetry.group.gcloud.dependencies]
google-cloud-storage = "^2.8.0"

[tool.poetry.scripts]
imgdl = 'imgdl.cli:main'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
